extracted complexgrid to its own file
