+++ src/components/GoogleMaps/GoogleMapComponent.js
@@ -1,22 +1,20 @@
 import React, {useState, useRef, useCallback} from "react";
 import { GoogleMap, useLoadScript, Marker, InfoWindow } from "@react-google-maps/api";
-import { Grid, Typography, IconButton, CardMedia , Paper, Box } from "@material-ui/core";
-import WifiIcon from '@material-ui/icons/Wifi';
-import NightsStayIcon from '@material-ui/icons/NightsStay';
-import Rating from '@material-ui/lab/Rating';
-import StarBorderIcon from '@material-ui/icons/StarBorder';
+import { Grid } from "@material-ui/core";
 import mapStyle from "./mapStyle";
 import places from "./places";
 import usePlacesAutoComplete , {getGeocode, getLatLng} from "use-places-autocomplete";
 import { makeStyles } from "@material-ui/core/styles";
 import { Combobox, ComboboxInput, ComboboxPopover, ComboboxList, ComboboxOption} from "@reach/combobox";
 import "@reach/combobox/styles.css";
+import {ComplexGrid}  from "./ComplexGrid";
 import campImageOne from "../../assets/coupleByCamper.jpg"
 import campImageTwo from "../../assets/camperInField.jpg"
 import campImageThree from "../../assets/camperUnderStars.jpg"
 import campImageFour from "../../assets/groupAtTable.jpg"
 import campImageFive from "../../assets/personOnRV.jpg"
-import campImageSix from "../../assets/sittingOnCamper.jpg"
+import campImageSix from "../../assets/sittingOnCamper.jpg";
+
 
 const useStyles = makeStyles((theme) => ({
   search: {
@@ -34,31 +32,6 @@ const useStyles = makeStyles((theme) => ({
     height: 48,
     zIndex: 10,
     border: "2px solid #39A5A7"
-  },
-  root: {
-    flexGrow: 1,
-    top: theme.spacing(10),
-    margin: theme.spacing(1),
-  },
-  paper: {
-    padding: theme.spacing(2),
-    maxWidth: 550,
-  },
-  image: {
-    width: 128,
-    height: 128,
-  },
-  img: {
-    margin: 'auto',
-    display: 'block',
-    maxWidth: '100%',
-    maxHeight: '100%',
-  },
-  cover: {
-    width: 160,
-    height:160,
-    padding: 0,
-    paddingRight: 2
   },
 }));
 
@@ -136,38 +109,19 @@ export const GoogleMapComponent = () => {
 
   return (
     <div>
-      <Grid container style={{ minHeight: "100vh" }}>
-        <Grid item xs={4}>
+      <Grid container style={{ maxHeight: "100vh" }}>
+        <Grid item lg={4} sm={4} xl={2} xs={0}>
           <div style={{ height: "80px" }} />
           {places.slice(0, 5).map((place) => (
             <div key={place.id}>
-              <ComplexGrid
-                imagePath={images[place.id - 1]}
-                campTitle={place.campgroundName}
-                campState={place.state}
-                campRating={place.rating}
-                campMaxDays={place.maxDays}
-              />
+              <ComplexGrid imagePath={images[place.id - 1]} campTitle={place.campgroundName} campState={place.state} campRating={place.rating} campMaxDays={place.maxDays}/>
             </div>
           ))}
         </Grid>
-        <Grid
-          item
-          container
-          xs={8}
-          alignContent="flex-end"
-          justify="center"
-          direction="column"
-        >
+        <Grid item lg={6} sm={8} xl={4} xs={4} container alignContent="flex-end" justify="center" direction="column">
           <Search panTo={panTo} />
           <Locate panTo={panTo} />
-          <GoogleMap
-            mapContainerStyle={containerStyle}
-            center={{ lat, lng }}
-            zoom={8}
-            options={options}
-            onLoad={onMapLoad}
-          >
+          <GoogleMap mapContainerStyle={containerStyle} center={{ lat, lng }} zoom={8} options={options} onLoad={onMapLoad}>
             {places.map((place) => (
               <Marker
                 key={place.id}
@@ -263,48 +217,4 @@ function Search({panTo}){
   );
 }
 
-function ComplexGrid({ imagePath, campTitle, campState, campRating, campMaxDays}) {
-  const classes = useStyles();
 
-  return (
-    <div className={classes.root}>
-      <Paper className={classes.paper} elevation={3}>
-        <Grid container spacing={4}>
-          <Grid item>
-          <CardMedia  className={classes.cover} image={imagePath} title="Live from space album cover" />
-          </Grid>
-          <Grid item xs={12} sm container>
-            <Grid item xs container direction="column" spacing={2}>
-              <Grid item xs>
-                <Typography gutterBottom variant="subtitle1">{campTitle}</Typography>
-                <Typography variant="body2" gutterBottom>{campState}</Typography>
-                <Box component="fieldset" mb={3} borderColor="transparent">
-                  <Typography component="legend">Overall Rating</Typography>
-                  <Rating
-                    name="customized-empty"
-                    defaultValue={campRating}
-                    precision={0.5}
-                    emptyIcon={<StarBorderIcon fontSize="inherit"
-                    size="large"
-                    readOnly />}
-                  />
-               </Box>
-              </Grid>
-            </Grid>
-            <Grid item>
-              <Grid item>
-                <IconButton >
-                  <NightsStayIcon style={{fontSize: 40, color: "#39A5A7"}}/>
-                  <Typography variant="subtitle1">Max {campMaxDays} </Typography>
-                </IconButton>
-              </Grid>
-                <IconButton >
-                  <WifiIcon style={{fontSize: 40, color: "#39A5A7"}}/>
-                </IconButton>
-              </Grid>
-          </Grid>
-        </Grid>
-      </Paper>
-    </div>
-  );
-}
